skip delete if query is empty, update tests

+++ src/core/packages/saved-objects/migration-server-internal/src/actions/cleanup_unknown_and_excluded.ts
@@ -21,7 +21,7 @@ import {
 } from './check_for_unknown_docs';
 import { isTypeof } from '.';
 import { CalculatedExcludeFilter, calculateExcludeFilters } from './calculate_exclude_filters';
-import { deleteByQuery } from './delete_by_query';
+import { deleteByQuery, DeleteByQueryResponse } from './delete_by_query';
 
 /** @internal */
 export interface CleanupUnknownAndExcludedParams {
@@ -46,6 +46,12 @@ export interface CleanupStarted {
   taskId: string;
 }
 
+export interface CleanupNotNeeded {
+  type: 'cleanup_not_needed';
+}
+
+type DeleteCleanupResult = CleanupNotNeeded | DeleteByQueryResponse;
+
 /**
  * Cleans up unknown and excluded types from the specified index.
  */
@@ -60,7 +66,7 @@ export const cleanupUnknownAndExcluded = ({
   removedTypes,
 }: CleanupUnknownAndExcludedParams): TaskEither.TaskEither<
   RetryableEsClientError | UnknownDocsFound,
-  CleanupStarted
+  CleanupStarted | CleanupNotNeeded
 > => {
   let unknownDocs: DocumentIdAndType[] = [];
   let unknownDocTypes: string[] = [];
@@ -92,43 +98,57 @@ export const cleanupUnknownAndExcluded = ({
     ),
 
     // actively delete unwanted documents
-    TaskEither.chainW((excludeFiltersRes) => {
-      errorsByType = excludeFiltersRes.errorsByType;
+    TaskEither.chainW(
+      (excludeFiltersRes): TaskEither.TaskEither<RetryableEsClientError, DeleteCleanupResult> => {
+        errorsByType = excludeFiltersRes.errorsByType;
 
-      // we must delete everything that matches:
-      // - any of the plugin-defined exclude filters
-      // - OR any of the unknown types
-      const deleteQuery: QueryDslQueryContainer = {
-        bool: {
-          should: [
-            ...excludeFiltersRes.filterClauses,
-            ...removedTypes.map((type) => ({ term: { type } })),
-            ...unknownDocTypes.map((type) => ({ term: { type } })),
-          ],
-        },
-      };
+        // we must delete everything that matches:
+        // - any of the plugin-defined exclude filters
+        // - OR any of the unknown types
+        const shouldClauses = [
+          ...excludeFiltersRes.filterClauses,
+          ...removedTypes.map((type) => ({ term: { type } })),
+          ...unknownDocTypes.map((type) => ({ term: { type } })),
+        ];
 
-      return deleteByQuery({
-        client,
-        indexName,
-        query: deleteQuery,
-        // we want to delete as many docs as we can in the current attempt
-        conflicts: 'proceed',
-        // instead of forcing refresh after each delete attempt,
-        // we opt for a delayRetry mechanism when conflicts appear,
-        // letting the periodic refresh kick in
-        refresh: false,
-      });
-    }),
+        if (shouldClauses.length === 0) {
+          return TaskEither.right({
+            type: 'cleanup_not_needed' as const,
+          } as CleanupNotNeeded);
+        }
+
+        const deleteQuery: QueryDslQueryContainer = {
+          bool: {
+            should: shouldClauses,
+          },
+        };
+
+        return deleteByQuery({
+          client,
+          indexName,
+          query: deleteQuery,
+          // we want to delete as many docs as we can in the current attempt
+          conflicts: 'proceed',
+          // instead of forcing refresh after each delete attempt,
+          // we opt for a delayRetry mechanism when conflicts appear,
+          // letting the periodic refresh kick in
+          refresh: false,
+        });
+      }
+    ),
 
     // map response output
-    TaskEither.chainEitherKW((res) => {
-      return Either.right({
-        type: 'cleanup_started' as const,
-        taskId: res.taskId,
-        unknownDocs,
-        errorsByType,
-      });
+    TaskEither.chainEitherKW((res): Either.Either<never, CleanupStarted | CleanupNotNeeded> => {
+      if (res.type === 'delete_by_query_response') {
+        return Either.right({
+          type: 'cleanup_started' as const,
+          taskId: res.taskId,
+          unknownDocs,
+          errorsByType,
+        });
+      } else {
+        return Either.right(res);
+      }
     })
   );
 };
