+++ src/core/packages/saved-objects/migration-server-internal/src/model/model.ts
@@ -537,27 +537,51 @@ export const model = (currentState: State, resW: ResponseType<AllActionStates>):
   } else if (stateP.controlState === 'CLEANUP_UNKNOWN_AND_EXCLUDED') {
     const res = resW as ExcludeRetryableEsError<ResponseType<typeof stateP.controlState>>;
     if (Either.isRight(res)) {
-      if (res.right.unknownDocs.length) {
+      if (res.right.type === 'cleanup_started') {
+        if (res.right.unknownDocs.length) {
+          logs = [
+            ...stateP.logs,
+            { level: 'warning', message: extractDiscardedUnknownDocs(res.right.unknownDocs) },
+          ];
+        }
+
         logs = [
-          ...stateP.logs,
-          { level: 'warning', message: extractDiscardedUnknownDocs(res.right.unknownDocs) },
+          ...logs,
+          ...Object.entries(res.right.errorsByType).map(([soType, error]) => ({
+            level: 'warning' as const,
+            message: `Ignored excludeOnUpgrade hook on type [${soType}] that failed with error: "${error.toString()}"`,
+          })),
         ];
+
+        return {
+          ...stateP,
+          logs,
+          controlState: 'CLEANUP_UNKNOWN_AND_EXCLUDED_WAIT_FOR_TASK',
+          deleteByQueryTaskId: res.right.taskId,
+        };
+      } else {
+        // cleanup_not_needed, let's move to the step after CLEANUP_UNKNOWN_AND_EXCLUDED_WAIT_FOR_TASK
+        const source = stateP.sourceIndex.value;
+        return {
+          ...stateP,
+          logs,
+          controlState: 'PREPARE_COMPATIBLE_MIGRATION',
+          targetIndexMappings: mergeMappingMeta(
+            stateP.targetIndexMappings,
+            stateP.sourceIndexMappings.value
+          ),
+          preTransformDocsActions: [
+            // Point the version alias to the source index. This let's other Kibana
+            // instances know that a migration for the current version is "done"
+            // even though we may be waiting for document transformations to finish.
+            { add: { index: source!, alias: stateP.versionAlias } },
+            ...buildRemoveAliasActions(source!, Object.keys(stateP.aliases), [
+              stateP.currentAlias,
+              stateP.versionAlias,
+            ]),
+          ],
+        };
       }
-
-      logs = [
-        ...logs,
-        ...Object.entries(res.right.errorsByType).map(([soType, error]) => ({
-          level: 'warning' as const,
-          message: `Ignored excludeOnUpgrade hook on type [${soType}] that failed with error: "${error.toString()}"`,
-        })),
-      ];
-
-      return {
-        ...stateP,
-        logs,
-        controlState: 'CLEANUP_UNKNOWN_AND_EXCLUDED_WAIT_FOR_TASK',
-        deleteByQueryTaskId: res.right.taskId,
-      };
     } else {
       const reason = extractUnknownDocFailureReason(
         stateP.migrationDocLinks.resolveMigrationFailures,
